feat: Completed the substrings project
This detects all of the substrings in the string provided by:
- Spliting the string on it's white space and finding if each word
 contains the substring. If so, iterate the counter.
- The program detects if the word contains a substring by applying a
\#sub method on the string with the substring as the parameter.
	- If the word contains the substring, then it's string will be replaced
	by an uppercased substring.
	- If the word is replaced, then it will no longer match the original word,
	thus we increment the counter.

To provide an example, here's the execution flow:
	string = "BeLow", dictionary = "low"

String gets downcased to ensure case-insensitiveness:
	"BeLow" => "below"

Since string contains "low" it becomes:
	"beLOW"

Thus:
	"below" != "beLOW"

So we increment the counter on that substring. Do these on each word
contained in the string and we get the final hash!